requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
requests
python-dotenv
httpx
python-multipart
rapidfuzz
